Advanced_Phase_Deliverables/Task_6_Quiz_Generator/requirements.txt
streamlit
pandas
langchain
langchain-google-genai
langchain-groq
python-dotenv
pypdf